requirements.txt
selenium==4.21.0
selenium-wire==5.1.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.2
